Update post-create.md
The example uses "*" for the "wildcard" specification, but the string "\*" is referenced in several places in the documentation for the service. I haven't verified this in actual code, but it seems very unlikely to me that the wildcard string would be "\*". I'm guessing it's more likely the example would have been tested, as opposed to the documentation.