Add link to Reference/Spin documentation
Google search results typically land on the JSON page, but there is no link to the XML and JSON spin data format reference documentation. This change adds the links so you do not have to re-search in the camunda docs for "json" in order to get to the spin reference docs.