Made "deployment aware" a seperate headline
This feature is very important and gains popularity in microservice scenarios, that's why it should have its own linkable headline. It is not really only about heterogeneous clusters. I also changed the headline for sharing the job executor when running multiple LOCAL engines - as this is not the default when running multiple engines and I even consider it quite a rare case.