Update serialization method used to a much simpler one.
While working in a forum ticket I realized that the code snippet to serialize the map object was a little off. Therefore I making a small recommendation so as to prevent tickets from being created (Not that it is a bad thing or anything :p). 

Reference to forum ticket: https://forum.camunda.org/t/how-to-load-default-values-in-a-task-form-from-h2-database/15178/4?u=patozgg